add static_handler.cc
